test: cover financial account routes
